Check if a version is rejected post-resolve
Previously, we checked if a version was rejected as part of selecting
that version in conflict resolution. This was problematic in a couple
of ways:
- The end result was order-dependent, since a failure in an early
  conflict resolution would abort resolve, preventing a later
  (non-rejected) version from being selected.
- There was no mechanism to do a "best effort" resolve, so we
  failed early. This prevented us from reporting all failures
  where multiple versions are rejected.

We now mark a chosen component as 'rejected' during conflict
resolution, but only fail the resolve later during a post-
validation phase. We still fail on the first discovered
error, but it should now be possible to collect all failures
for reporting.